fix(prose/golang): accept purpose of use via config for now
For some reason peer metadata is not included with each request, so we
cannot reliably extract purpose of use from there. Before we figure it
out, we are going to accept it as a configuration option. We are also
injecting this configuration option into each of the filters associated
with each service. Instead of copying the filter source files a few
times for each of the deployments, we are generating a few different
instances of the same filter with different names and selectors based on
the postBuild substitution feature of flux kustomizations.

Fixes #101