feat(eval/latencies): inspect prose-filter without presidio
We want to inspect the behaviour of prose-filter that sleeps 20ms
instead of calling real presidio server. The change was introduced in
8ec667ab332b140000b15a43fc84e8d75c2fc47f.

We also do not want to attack presidio during this test. So this commit
also reverts e0f8c28a0b52755d6ca0d41fba1af76b04e0420b and
7619edb91537ee1c54bbd56a5059ffa23b2d09e8.